chore: split Filter into different modules
