+++ packages/sdk/src/index.ts
@@ -13,7 +13,7 @@ export * from "./waku.js";
 
 export { createLightNode, createNode } from "./light-node/index.js";
 export { wakuLightPush } from "./protocols/light_push.js";
-export { wakuFilter } from "./protocols/filter.js";
+export { wakuFilter } from "./protocols/filter/index.js";
 export { wakuStore } from "./protocols/store.js";
 
 export * as waku from "@waku/core";
