+++ packages/sdk/src/relay-node/index.ts
@@ -1,7 +1,7 @@
 import { type FullNode, type RelayNode } from "@waku/interfaces";
 import { RelayCreateOptions, wakuRelay } from "@waku/relay";
 
-import { wakuFilter } from "../protocols/filter.js";
+import { wakuFilter } from "../protocols/filter/index.js";
 import { wakuLightPush } from "../protocols/light_push.js";
 import { wakuStore } from "../protocols/store.js";
 import { createLibp2pAndUpdateOptions } from "../utils/libp2p.js";
